OFX is not relevant to model.
